// src/interfaces/notification-layout.interface.ts
interface INotificationLayout {
  width: number;
  centerX: number;
  islandWidth: number;
  islandHeight: number;
  islandTop: number;
  islandBottom: number;
  islandCenterY: number;
  islandRadius: number;
  cardWidth: number;
  cardHeight: number;
  cardRadius: number;
  cardTop: number;
  cardLeft: number;
  cardCenterY: number;
  dropSize: number;
  neckWidth: number;
  canvasHeight: number;
}

export type { INotificationLayout };